Ciclistas iniciam sua propria posiçao em metros d v e n sao globais

+++ EP2/ciclista.h
@@ -8,7 +8,7 @@ typedef struct {
 	int volta; /* Volta atual do ciclista */
 	int p; /* Pontuação acumulada do ciclista */
 	int v, vMax; /* Velocidade desejada e máxima do ciclista */
-	float pos; /* Posição em metros na pista */
+	int pos; /* Posição em metros na pista */
 	int quebrado; /* Guarda a volta em que ciclista quebrou. Se não, -1 */
 } ciclista;
